+++ arrow/src/main/java/org/apache/iceberg/arrow/vectorized/parquet/VectorizedRleBooleanValuesReader.java
@@ -19,15 +19,8 @@
 package org.apache.iceberg.arrow.vectorized.parquet;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import org.apache.arrow.vector.BitVector;
 import org.apache.arrow.vector.FieldVector;
-import org.apache.parquet.bytes.ByteBufferInputStream;
-import org.apache.parquet.bytes.BytesUtils;
-import org.apache.parquet.column.values.ValuesReader;
-import org.apache.parquet.column.values.bitpacking.BytePacker;
-import org.apache.parquet.column.values.bitpacking.Packer;
-import org.apache.parquet.io.ParquetDecodingException;
 
 /**
  * A {@link VectorizedValuesReader} for RLE-encoded boolean data pages.
@@ -36,34 +29,18 @@ import org.apache.parquet.io.ParquetDecodingException;
  *     href="https://parquet.apache.org/docs/file-format/data-pages/encodings/#run-length-encoding--bit-packing-hybrid-rle--3">
  *     Parquet format encodings: RLE</a>
  */
-public class VectorizedRleBooleanValuesReader extends ValuesReader
+public class VectorizedRleBooleanValuesReader extends VectorizedRLEAndBitPackingParquetValuesReader
     implements VectorizedValuesReader {
 
-  private enum Mode {
-    RLE,
-    PACKED
-  }
-
-  private static final int BIT_WIDTH = 1;
-  private static final BytePacker PACKER = Packer.LITTLE_ENDIAN.newBytePacker(BIT_WIDTH);
-
-  private ByteBufferInputStream inputStream;
-  private Mode mode;
-  private int currentCount;
-  private int currentValue;
-  private int[] packedValuesBuffer = new int[16];
-  private int packedValuesBufferIdx = 0;
-
   @Override
-  public void initFromPage(int valueCount, ByteBufferInputStream in) throws IOException {
-    int length = BytesUtils.readIntLittleEndian(in);
-    this.inputStream = in.sliceStream(length);
-    this.currentCount = 0;
+  int encodedDataLength() throws IOException {
+    init(1);
+    return readIntLittleEndian();
   }
 
   @Override
-  public boolean readBoolean() {
-    return nextValue() != 0;
+  int readIntLittleEndianPaddedOnBitWidth() throws IOException {
+    return inputStream.read();
   }
 
   @Override
@@ -72,68 +49,5 @@ public class VectorizedRleBooleanValuesReader extends ValuesReader
     for (int i = 0; i < total; i++) {
       bitVector.setSafe(rowId + i, readBoolean() ? 1 : 0);
     }
-  }
-
-  private int nextValue() {
-    if (currentCount == 0) {
-      readNextGroup();
-    }
-
-    currentCount--;
-    switch (mode) {
-      case RLE:
-        return currentValue;
-      case PACKED:
-        return packedValuesBuffer[packedValuesBufferIdx++];
-    }
-    throw new RuntimeException("Unrecognized mode: " + mode);
-  }
-
-  private void readNextGroup() {
-    try {
-      int header = readUnsignedVarInt();
-      this.mode = (header & 1) == 0 ? Mode.RLE : Mode.PACKED;
-      switch (mode) {
-        case RLE:
-          this.currentCount = header >>> 1;
-          this.currentValue = inputStream.read();
-          return;
-        case PACKED:
-          int numGroups = header >>> 1;
-          this.currentCount = numGroups * 8;
-          if (this.packedValuesBuffer.length < this.currentCount) {
-            this.packedValuesBuffer = new int[this.currentCount];
-          }
-          packedValuesBufferIdx = 0;
-          int valueIndex = 0;
-          while (valueIndex < this.currentCount) {
-            ByteBuffer buffer = inputStream.slice(BIT_WIDTH);
-            PACKER.unpack8Values(buffer, buffer.position(), this.packedValuesBuffer, valueIndex);
-            valueIndex += 8;
-          }
-          return;
-        default:
-          throw new ParquetDecodingException("not a valid mode " + this.mode);
-      }
-    } catch (IOException e) {
-      throw new ParquetDecodingException("Failed to read from input stream", e);
-    }
-  }
-
-  private int readUnsignedVarInt() throws IOException {
-    int value = 0;
-    int shift = 0;
-    int byteRead;
-    do {
-      byteRead = inputStream.read();
-      value |= (byteRead & 0x7F) << shift;
-      shift += 7;
-    } while ((byteRead & 0x80) != 0);
-    return value;
-  }
-
-  @Override
-  public void skip() {
-    throw new UnsupportedOperationException();
   }
 }
